Updating

+++ ROUTER/category.py
@@ -9,13 +9,13 @@ router = APIRouter(prefix="/category", tags=["Categories"])
 @router.post("/create")
 def create_category(category: schemas.CategoryCreated, db: Session = Depends(database.get_db), 
                     current_user = Depends(oauth2.get_current_user)):
-    existing_category = db.query(models.NoteCategory).filter_by(category_name= category.category_name.title(), 
+    existing_category = db.query(models.NoteCategory).filter_by(category_name= category.category_name, 
                                                                 user_id = current_user.id).first()
     if existing_category:
         raise HTTPException(status_code=status.HTTP_409_CONFLICT, 
                             detail="Category already exists")
     new_category = models.NoteCategory(user_id=current_user.id, 
-                                       category_name=category.category_name.title())
+                                       category_name=category.category_name)
     db.add(new_category)
     db.commit()
     db.refresh(new_category)
@@ -42,32 +42,35 @@ def get_category_note(category_id: int, db: Session = Depends(database.get_db),
 
 
 @router.put("/edit/{category_id}")
-def edit_category(category_id: int, category: schemas.EditCategory, 
-                  db: Session = Depends(database.get_db), 
+def edit_category(category_id: int, category: schemas.EditCategory, db: Session = Depends(database.get_db), 
                   current_user = Depends(oauth2.get_current_user)):
-    category_query = db.query(models.NoteCategory).filter_by(id = category_id, 
-                                                             user_id = current_user.id)
+
+    category_query = db.query(models.NoteCategory).filter_by(id=category_id, user_id=current_user.id)
     editing_category = category_query.first()
     if not editing_category:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
-                            detail="Category doesn't exist")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Category doesn't exist")
+    
+    if editing_category.category_name == "Uncategorized":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Cannot edit default category")
 
     category_query.update(category.model_dump(), synchronize_session=False)
     db.commit()
     edited_category = category_query.first()
     return edited_category
 
+
 @router.delete("/delete/{category_id}", response_model=schemas.Deletion)
 def delete_category(category_id:int, db: Session = Depends(database.get_db), 
-                  current_user = Depends(oauth2.get_current_user)):
-    
-    category_query = db.query(models.NoteCategory).filter_by(id=category_id, 
-                                                             user_id=current_user.id)
+                    current_user = Depends(oauth2.get_current_user)):
+
+    category_query = db.query(models.NoteCategory).filter_by(id=category_id, user_id=current_user.id)
     category = category_query.first()
     
     if not category:
-       raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
-                           detail=f"Category doesn't exist")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Category doesn't exist")
+
+    if category.category_name == "Uncategorized":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Cannot delete default category")
    
     category_query.delete(synchronize_session=False)
     db.commit()
